refactor: Query 모듈에 관련된 로직 인프라 모듈로 이동
[#64]
