+++ build.gradle.kts
@@ -76,24 +76,3 @@ project(":command") {
         implementation(project(":domain:mysql"))
     }
 }
-
-project(":query") {
-    dependencies {
-        implementation(project(":common"))
-        implementation(project(":domain:mongo"))
-        implementation(project(":infrastructure:redis"))
-        implementation(project(":infrastructure:restclient"))
-    }
-}
-
-project(":infrastructure:redis") {
-    dependencies {
-        implementation(project(":common"))
-    }
-}
-
-project(":infrastructure:restclient") {
-    dependencies {
-        implementation(project(":common"))
-    }
-}
