--- infrastructure/mongo/build.gradle.kts
@@ -1,12 +1,0 @@
-plugins {
-    id("org.springframework.boot")
-
-    kotlin("jvm")
-    kotlin("plugin.spring")
-    kotlin("plugin.allopen")
-    kotlin("plugin.jpa")
-    kotlin("kapt")
-}
-
-dependencies {
-}
