+++ infrastructure/restclient/build.gradle.kts
@@ -1,18 +1,9 @@
-import org.springframework.boot.gradle.tasks.bundling.BootJar
-
-val jar: Jar by tasks
-val bootJar: BootJar by tasks
-
-bootJar.enabled = false
-jar.enabled = true
-
 plugins {
-    id("org.springframework.boot")
-
-    kotlin("jvm")
     kotlin("plugin.spring")
 }
 
 dependencies {
+    implementation(project(":common"))
+
     implementation("org.springframework.boot:spring-boot-starter-web")
 }
