+++ query/build.gradle.kts
@@ -3,18 +3,16 @@ plugins {
 
     kotlin("jvm")
     kotlin("plugin.spring")
-    kotlin("plugin.allopen")
-    kotlin("plugin.jpa")
-}
-
-allOpen {
-    // MongoDB Document 키워드 Open
-    annotation("org.springframework.data.mongodb.core.mapping.Document")
 }
 
 dependencies {
+    implementation(project(":common"))
+    implementation(project(":domain:mongo"))
+    implementation(project(":infrastructure:mongodb"))
+    implementation(project(":infrastructure:redis"))
+    implementation(project(":infrastructure:restclient"))
+
     implementation("org.springframework.boot:spring-boot-starter-amqp")
-    implementation("org.springframework.boot:spring-boot-starter-data-mongodb")
     implementation("org.springframework.boot:spring-boot-starter-validation")
     implementation("org.springframework.boot:spring-boot-starter-web")
 }
